refactor: 拆分 mcp_server 为工具目录与参数映射模块
将 1624 行的 mcp_server.py 拆成 mcp_tools（TOOLS/合规过滤）与
mcp_args（_build_args），主模块只保留服务器与传输层；公开符号原样
再导出，wrapper 与测试导入路径不变。同步将新模块纳入 mypy 严格白名单。
